fix rounding
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/pathsoft/artemis/trunk@11955 ee4ac58c-ac51-4696-9907-e4b3aa274f04

+++ uk/ac/sanger/artemis/io/StreamSequence.java
@@ -129,7 +129,7 @@ public abstract class StreamSequence
 
 //  int packStart = Math.round( (float)start/2.f ) - 1;
     int packStart = (start - 1) >> 1;
-    int packEnd   = Math.round( (float)end/2.f ); // end/2;
+    int packEnd   = (int) Math.round( (double)end/2.d ); // end/2;
     
     int count = 0;
     byte currStorageUnit;
